api-watcher-controller/requirements.txt
kubernetes
pick
requests
python-socketio
asyncio
aiohttp